requirements.txt
kafka-python
pyspark
diagrams
graphviz